backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
python-dotenv

# Auth / JWT
python-jose

# Supabase
supabase

# Google Gemini (new SDK)
google-genai

# Optional but useful
typing-extensions
